% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tLagInterimData.R
\name{tLagIntCont}
\alias{tLagIntCont}
\alias{b.data.cont}
\alias{x.data.cont}
\alias{t.data.cont}
\title{Toy Dataset With a Continuous Outcome For Illustration}
\format{
The dataset provides three data.frames: \code{b.data.cont} containing the
basic observed data on 245 enrolled subjects at the time of an interim
analysis, with columns with headers
\itemize{
\item "subjID" - unique subject identifiers,
\item "u" - minimum of time lag or censoring time,
\item "delta" - time lag/censoring indicator, and
\item "Y" - the outcome if it is available, = 0 if not.
\item "a" - treatment indicator;
}

\code{x.data.cont} contains the baseline covariates for the 245 subjects.
\itemize{
\item "subjID" - unique subject identifiers,
\item "X1" - a continuous covariate;
}

and \code{t.data.cont} contains time-dependent covariate information comprising
"subjID" and 6 measurements of a single continuous covariate measured
at times (t1 = 0, t2 = 4, t3 = 12, t4 = 24, t5 = 52)
}
\usage{
data(tLagIntCont)
}
\description{
These data are provided for the purposes of illustrating the use of
the software when the outcome of interest is continuous.
Though these data were generated to mimic conditions of a clinical trial,
they should not be interpreted as representing true clinical trial data.
}
\keyword{datasets}
